\name{Ants}
\Rdversion{2.1}
\alias{Ants}
\docType{data}
\title{
Movements of ants
}
\description{Directions chosen by 100 ants in response to an evenly illuminated
black target.
}
\usage{data(Ants)}
\format{Directions chosen by 100 ants in degrees}
\source{
Randomly selected values by Fisher (1993) from Jander (1957)}
\references{
N.I. Fisher (1993) Statistical analysis of circular data. Cambridge University Press.}
%\examples{
%data(Ants)
%## maybe str(Ants) ; plot(Ants) ...
%}
\keyword{datasets}
