\name{Ants_radians}
\Rdversion{1.0-1}
\alias{Ants_radians}
\docType{data}
\title{
Movements of ants
}
\description{
Direction chosen by 100 ants in response to an evenly illuminated
black target.
}
\usage{data(Ants_radians)}
\format{Directions chosen by 100 ants in radians}

\source{
Randomly selected values by Fisher (1993) from Jander (1957)
}
\references{
N.I. Fisher (1993) Statistical analysis of circular data. Cambridge University Press.
}

\keyword{datasets}
