\name{Datab3fisher}
\Rdversion{2.1}
\alias{Datab3fisher}
\docType{data}
\title{Database B3 from Fisher}
\description{
Database B3 from Fisher et al. (1987)
}
\usage{data(Datab3fisher)}
\format{Datab3fisher}
\details{The dataset Datab3fisher consists of 148 observations of the arrival
directions of low-mu showers of cosmic rays (Toyoda \emph{et al.}, 1965; see Fisher \emph{et
al.}, 1987, pp. 280-281).
The observations are measured in declination and right ascension coordinates.
}
\source{
Fisher, et al. (1987)
}
\references{Toyoda, Y., Suga, K., Murakami, K., Hasegawa, H.,
Shibata, S., Domingo, V., Escobar, I., Kamata, K., Bradt, H.,
Clark, G. and La Pointe, M. (1965). Studies of Primary Cosmic Rays
in the Energy Region $10^{14}$ eV to $10^{17}$ eV (Bolivian Air
Shower Joint Experiment), Proceedings of the International
Conference on Cosmic Rays, vol. 2,
London, September, 1965, 708--711. London: The Institute of
Physics and the Physical Society.

Fisher, N.I., Lewis, T. and Embleton, B.J.J. (1987).
Statistical Analysis of Spherical Data,
Cambridge U.K.: Cambridge University Press.
}
%\examples{
%data(datab6fisher)
%## maybe str(datab6fisher) ; plot(datab6fisher) ...
%}
\keyword{datasets}
