\name{DataB5FisherSpherical}
\Rdversion{2.1}
\alias{DataB5FisherSpherical}
\docType{data}
\title{Spherical Data on Magnetic Remanence }
\description{Measurements of magnetic remanence from 52 specimens of red beds from the Bowen Basin, Queensland.}
\usage{data(DataB5FisherSpherical)}
\format{Declination -inclination in degrees}
\source{P.W. Schmidt  }
\references{
Fisher N.I., Lewis T. and Embleton B.J.J. (1987) Statistical Analysis of Spherical Data.
Cambridge University Press, Cambridge.
Data B.5.  }
\examples{
data(DataB5FisherSpherical)
%## maybe str() ; plot() ...
}
\keyword{datasets}
