\name{Datab6fisher}
\Rdversion{2.1}
\alias{Datab6fisher}
\docType{data}
\title{Database B6 from Fisher et al. (1987)}
\description{
datab6fisher
}
\usage{data(Datab6fisher)}
\format{The coordinates are declination and inclination measured in degrees}
\details{The data-set Datab6fisher contains 107 measurements of magnetic remanence
in samples of Precambrian volcanics collected in Northwest Australia.
(Schmidt and Embleton, 1985; see Fisher et al., 1987, pp. 285).
}
\source{
Fisher, et al. (1987)
}
\references{Schmidt, P.W. and Embleton, B.J.J. (1985).
Pre-folding and overprint magnetic signatures in
Precambrian (~2.9-2.7ga) igneous rocks from the Pilbara Craton and
Hamersley Basin, N.W. Australia,
Journal of Geophysical Research, 90 (B4), 2967--2984.

Fisher, N.I., Lewis, T. and Embleton, B.J.J. (1987).
Statistical Analysis of Spherical Data,
Cambridge U.K.: Cambridge University Press.
}
%\examples{
%data(datab6fisher)
%## maybe str(datab6fisher) ; plot(datab6fisher) ...
%}
\keyword{datasets}
