\name{DataUniformBivariate200obs}
\Rdversion{2.1}
\alias{DataUniformBivariate200obs}
\docType{data}
\title{Uniform Bivariate Circular Data}
\description{200 realizations of a uniform distribution on the torus }
\usage{data(DataUniformBivariate200obs)}
\format{Angles in radians }
%\examples{
%data( )
%## maybe str() ; plot() ...
%}
\keyword{datasets}
