\name{EarthquakesPacificMexicogt6}
\Rdversion{2.1}
\alias{EarthquakesPacificMexicogt6}
\docType{data}
\title{Date of Occurrence of Earthquakes}
\description{The time of occurrence of earthquakes of intensity greater than 6.0$^o$ Richter with an epicenter occurring in the coast
of the Pacific Ocean in Mexico from 1920 to 2002.
There is a total of 241 observations.}
\usage{data(EarthquakesPacificMexicogt6)}
\format{Time in years. All observations in the interval (0,1] }
\source{Mexican Database of Strong Earthquakes. CENAPRED. }
%\examples{
%data( )
%## maybe str() ; plot() ...
%}
\keyword{datasets}
