\name{EarthquakesPacificMexicogt7}
\Rdversion{2.1}
\alias{EarthquakesPacificMexicogt7}
\docType{data}
\title{Date of Occurrence of Earthquakes 2}
\description{The time of occurrence of earthquakes of intensity greater than \eqn{7.0^o} Richter with an epicenter occurring in the coast
of the Pacific Ocean in Mexico from 1920 to 2002.
There are a total of 76 observations. }
\usage{data(EarthquakesPacificMexicogt7)}
\format{Time in years. All observations in the interval (0,1] }
\source{Mexican Database of Strong Earthquakes. CENAPRED. }
%\examples{
%data( )
%## maybe str() ; plot() ...
%}
\keyword{datasets}
