\name{HomicidesMexico2005}
\Rdversion{2.1}
\alias{HomicidesMexico2005}
\docType{data}
\title{Homicides in Mexico during 2005}
\description{Monthly number of homicides in Mexico during 2005 }
\usage{data(HomicidesMexico2005)}
\format{Integer values }
\source{INEGI (Mexican National Statistical Agency)
www.inegi.gob.mx }
%\examples{
%data( )
%## maybe str() ; plot() ...
%}
\keyword{datasets}
