\name{HurricanesGulfofMexico1951to1970}
\Rdversion{2.1}
\alias{HurricanesGulfofMexico1951to1970}
\docType{data}
\title{Hurricanes in Mexico from 1951 to 1970}
\description{The time of occurrence (starting times) of hurricanes in the Gulf of Mexico for the 1951-1970 period.
There are a total of 196 observations. }
\usage{data(HurricanesGulfofMexico1951to1970)}
\format{Time in years. All observations in the interval (0,1] }
\source{http://weather.unisys.com/hurricane/atlantic/1978/index.html }
%\examples{
%data( )
%## maybe str() ; plot() ...
%}
\keyword{datasets}
