\name{HurricanesGulfofMexico1971to2008}
\Rdversion{2.1}
\alias{HurricanesGulfofMexico1971to2008}
\docType{data}
\title{Hurricanes in Mexico from 1971 to 2008}
\description{The time of occurrence (starting times) of hurricanes in the Gulf of Mexico for the 1971-2008 period.
There are a total of 417 observations }
\usage{data(HurricanesGulfofMexico1971to2008)}
\format{Time in years. All observations in the interval (0,1] }
\source{http://weather.unisys.com/hurricane/atlantic/1978/index.html }
%\examples{
%data( )
%## maybe str() ; plot() ...
%}
\keyword{datasets}
