\name{nntsABcoefficientsSymmetric}
\Rdversion{1.0}
\alias{nntsABcoefficientsSymmetric}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{AB coefficients}
\description{This function transforms the complex parameters c to the parameters ab for a
reparameterization of the density function}
\usage{nntsABcoefficientsSymmetric(cpars = c(0,0), M = 0)}
\arguments{
\item{cpars}{Vector of complex numbers of dimension 2*M}
\item{M}{Number of components in the NNTS}}
\value{The function returns the parameters ab associated with the pararameters cpars}
%%\references{
%% ~put references to the literature/web site here ~
%%}
\author{Juan Jose Fernandez-Duran and Maria Mercedes Gregorio-Dominguez}
%%\note{This function is required for compatibility in the Symmetric case
%%  ~~further notes~~
%%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

%%\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
%%}
%\examples{
%}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
