\name{ProteinsAAA}
\Rdversion{2.1}
\alias{ProteinsAAA}
\docType{data}
\title{Dihedral angles in protein}
\description{Dataset of the dihedral angles in a protein between three consecutive Alanine (Ala) amino acids. This dataset was
constructed from the recommended July 2003 list of proteins via
the algorithm in Hobohm et al. (1992). This algorithm selects a representative sample of proteins
from the vast Protein Data Bank (PDB, Berman et al., 2000). The dataset contains 233 pairs of dihedral angles.}
\usage{data(ProteinsAAA)}
\format{Two columns of angles in radians }
\source{Protein Data Bank (PDB) }
\references{Hobohm, U. and Scharf, M. and Schneider, R. and Sander, C. (1992) Selection of a Representative Set of Structures from the Brookhaven Protein Data Bank,
Protein Science, 1, 409-417.
Berman, H. M. and Westbrook, J. and Feng, Z. and Gilliand, G. and Bhat, T. N. and Weissing, H. and Shyndialov, I. N. and Bourne, P. E. (2000) 
The Protein Data Bank, Nucleic Acids Research, 28, 235-242.
 }
%\examples{
%data( )
%## maybe str() ; plot() ...
%}
\keyword{datasets}
