\name{SuicidesMexico2005}
\Rdversion{2.1}
\alias{SuicidesMexico2005}
\docType{data}
\title{Suicides in Mexico during 2005}
\description{Monthly number of suicides in Mexico during 2005 }
\usage{data(SuicidesMexico2005)}
\format{Integer values }
\source{ INEGI (Mexican National Statistical Agency)
www.inegi.gob.mx}
%\examples{
%data( )
%## maybe str() ; plot() ...
%}
\keyword{datasets}
