\name{turtles_radians}
\Rdversion{0.1}
\alias{turtles_radians}
\docType{data}
\title{
Movements of turtles
}
\description{
Data measurement of the directions taken by 76 turtles after treatment
}
\usage{data(turtles_radians)}
\format{Directions of turtles in radians}

%\details{
%}
\source{
Stephens (1969) Techniques for directional data. Technical Report 150. Dept.
of Statistics, Stanford University. Stanford, CA.
}
\references{
N.I. Fisher (1993) Statistical analysis of circular data. Cambridge University Press.
}
%\examples{
%data(turtles_radianes)
%## maybe str(turtles_radianes) ; plot(turtles_radianes) ...
%}
%\keyword{datasets}
