\name{ants}
\Rdversion{0.1}
\alias{ants}
\docType{data}
\title{
Movements of ants
}
\description{Directions chosen by 100 ants in response to an evenly illuminated
black target.
}
\usage{data(ants)}
\format{Directions chosen by 100 ants in degrees}
\source{
Randomly selected values by Fisher (1993) from Jander (1957)}
\references{
N.I. Fisher (1993) Statistical analysis of circular data. Cambridge University Press.}
%\examples{
%data(ants)
%## maybe str(ants) ; plot(ants) ...
%}
%\keyword{datasets}
