\name{Datab3fisher_ready}
\Rdversion{2.0}
\alias{Datab3fisher_ready}
\docType{data}
\title{Data transformed form Datab3fisher}
\description{Data transformed form Datab3fisher}
\usage{data(Datab3fisher_ready)}
\format{Datab3fisher_ready}
\details{
datab3fisher[,2] <- 90 + datab3fisher[,2];
datab3fisher_ready <- datab3fisher*(pi/180)}
%\source{
%Fisher (1993)
%}

%\examples{
%data(datab6fisher_ready)
%## maybe str(datab6fisher_ready) ; plot(datab6fisher_ready) ...
%}
\keyword{datasets}
